refactor: interval synthetic and organic query runner to separate class
